Восстановление бранча после внезапного удаления (добавил фукнционал уведомлений и еще испрвлены некоторые проблемы в редакторе)

+++ portal/announcements/management/commands/delete_expired_values.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 from django.core.management.base import BaseCommand
-from announcements.models import Announcement
+from announcements.models import Announcement, Notification
 
 
 class Command(BaseCommand):
@@ -11,13 +11,16 @@ class Command(BaseCommand):
         announcements = Announcement.objects.all()
 
         for announcement in announcements:
-            if announcement.date_of_expiring <= now:
+            if announcement.date_of_expiring:
+                if announcement.date_of_expiring <= now:
 
-                files = announcement.files.all()
+                    files = announcement.files.all()
+                    notifications = Notification.objects.filter(announcement_id=announcement.id)
 
-                for file in files:
-                    file.file.delete()
-                    file.delete()
+                    for file in files:
+                        file.file.delete()
+                        file.delete()
 
-                self.stdout.write(self.style.SUCCESS(f'Successfully deleted announcement with id {announcement.id} and its related image and files.'))
-                announcement.delete()
+                    self.stdout.write(self.style.SUCCESS(f'Successfully deleted announcement with id {announcement.id} and its related image and files.'))
+                    announcement.delete()
+                    notifications.delete()
